Refactor batch calls into pipeline.
