+++ consultation_analyser/consultations/views/pages.py
@@ -2,8 +2,8 @@ from django.http import HttpRequest
 from django.shortcuts import render
 from waffle.decorators import waffle_switch
 
-from consultation_analyser.batch_calls import BatchJobHandler
 from consultation_analyser.hosting_environment import HostingEnvironment
+from consultation_analyser.pipeline.batch_calls import BatchJobHandler
 
 from .. import models
