+++ tests/unit/test_batch_calls.py
@@ -3,7 +3,7 @@ from unittest.mock import MagicMock, patch
 
 from django.conf import settings
 
-from consultation_analyser.batch_calls import BatchJobHandler
+from consultation_analyser.pipeline.batch_calls import BatchJobHandler
 
 
 @patch("boto3.client")
